tcltk-package {tcltk}	R Documentation

Tcl/Tk Interface

Description

Interface and language bindings to Tcl/Tk GUI elements.

Details

This package provides access to the platform-independent Tcl scripting language and Tk GUI elements. See TkWidgets for a list of supported widgets, TkWidgetcmds for commands to work with them, and references in those files for more.

The Tcl/Tk documentation is in the system man pages.

For a complete list of functions, use ls("package:tcltk").

Note that Tk will not be initialized if there is no DISPLAY variable set, but Tcl can still be used. This is most useful to allow the loading of a package which depends on tcltk in a session that does not actually use it (e.g., during installation).

Author(s)

R Core Team

Maintainer: R Core Team R-core@r-project.org

[Package tcltk version 3.2.0 Index]