tkStartGUI {tcltk}	R Documentation

Tcl/Tk GUI startup

Description

Starts up the Tcl/Tk GUI

Usage


tkStartGUI()

Details

Starts a GUI console implemented via a Tk text widget. This should probably be called at most once per session. Also redefines the file pager (as used by help()) to be the Tk pager.

Note

This function is not usable under Windows.

tkStartGUI() saves its evaluation environment as .GUIenv. This means that the user interface elements can be accessed in order to extend the interface. The three main objects are named Term, Menu, and Toolbar, and the various submenus and callback functions can be seen with ls(envir = .GUIenv).

Author(s)

Peter Dalgaard

[Package tcltk version 3.6.1 Index]