tk_messageBox {tcltk}	R Documentation

Tk Message Box

Description

An implementation of a generic message box using Tk.

Usage


tk_messageBox(type = c("ok", "okcancel", "yesno", "yesnocancel",
                       "retrycancel", "abortretryignore"),
              message, caption = "", default = "", ...)

Arguments

type
character. The type of dialog box. It will have the buttons implied by its name. Can be abbreviated.

message
character. The information field of the dialog box.

caption
the caption on the widget displayed.

default
character. The name of the button to be used as the default.

...
additional named arguments to be passed to the Tk function of this name. An example is icon = "warning".

Value

A character string giving the name of the button pressed.

See Also

tkmessageBox for a ‘raw’ interface.

[Package tcltk version 3.6.1 Index]