tcltk-defunct {tcltk}	R Documentation

Defunct Functions in Package tcltk

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage

tkcmd(...)

tkfile.tail(...)
tkfile.dir(...)
tkopen(...)
tkclose(...)
tkputs(...)
tkread(...)

Arguments

... 	 Handled via .Tcl.args

Details

At R 2.0.0, tkcmd was renamed to tcl. The file commands used to be tkopen, tkclose, etc., but since open and close do not belong to the Tk toolkit, the preferred names are now tclopen, etc.

The $ methods for class "tclvar" have been deprecated in favour of tclVar and tclvalue.

See Also

Defunct

[Package tcltk version 2.9.2 Index]