tkpager {tcltk}	R Documentation

Page file using Tk text widget

Description

This plugs into file.show, showing files in separate windows.

Usage

tkpager(file, header, title, delete.file)

Arguments

file 	 character vector containing the names of the files to be displayed

header 	 headers to use for each file

title 	 common title to use for the window(s). Pasted together with the header to form actual window title.

delete.file 	 logical. Should file(s) be deleted after display?

Note

The "\b_" string used for underlining is currently quietly removed. The font and background colour are currently hardcoded to Courier and gray90.

See Also

file.show

[Package tcltk version 2.9.2 Index]